linux-tegra: add bbappend with mender workaround
The linux-%.bbappend in meta-mender-core was adding
itself to FILESEXTRAPATHS and possibly altering the
kernel config.  Tegra platforms don't need this config,
and the addition to FILESEXTRAPATHS was changing the
hash for the do_kernel_metadata task (which checksums
all config fragment files in all paths, used or not).

So to enable sharing of kernels between mender and
non-mender builds, undo the uneeded changes.

Signed-off-by: Matt Madison <matt@madison.systems>