systemd: disable wheel-group
By default, systemd will assume the 'wheel' group is available
(or can be created via the sysusers mechanism) and adds ACLs
that mention this group.  This leads to warnings during startup
about the non-existent group when systemd-sysusers can't create
it, or worse, having systemd-sysusers create the group when
we don't want it to.

Signed-off-by: Matt Madison <matt@madison.systems>