td-initramfs: trim fstab, remove systemd-extra-utils
We can defer the mounting of the added partitions to
after the rootfs has been mounted, so trim off everything
after the /run entry.  This also simplifies builds for
secureboot platforms, since we can use the same initramfs
for both the installer and normal images with this setup.

Also remove systemd-extra-utils, which is RRECOMMENDed by
systemd, but not really needed in the initramfs.

Signed-off-by: Matt Madison <matt@madison.systems>