testdistro-mender.conf: fix Mender breakage of DISTRO_FEATURES_BACKFILL
Mender adds 'systemd' to DISTRO_FEATURES_BACKFILL, rather
than to DISTRO_FEATURES, when the 'mender-systemd' feature
is enabled. By setting it as a 'backfill' feature, the change
also affects the DISTRO_FEATURES computation for non-target
recipes. In particular, it adds 'systemd-systemctl-native'
as a dependency for all -native recipes that inherit systemd,
which has ripple effects on task signatures throughout the
build, and makes it impossible to share sstate between
Mender and non-Mender builds.

Signed-off-by: Matt Madison <matt@madison.systems>