tegra-sysinstall: switch to installer-specific kernel+initrd
The installer image never has encrypted partitions, but sets them
up for use by regular images.  It will be simpler to manage this
flow to keep the initramfs separate.

Also remove mender-client (and its dependencies) from the
installer, since it is not used there but is slipped in
by mender-setup under the hood.

Signed-off-by: Matt Madison <matt@madison.systems>