mender-custom-flash-layout: fix Nano layout, add one for Nano-2GB
The SPI flash layout needed fixing for the 4GB Nano due to
some changes in the L4T flashing tools.  The 2GB Nano uses
an identical layout for now, but may adjust it later to
add a swap partition, since swap use is recommended by
NV on the 2GB model.

Signed-off-by: Matt Madison <matt@madison.systems>