u-boot-tegra: update for config fragment use with Mender
In our dual-rootfs flash layouts that are shared between
the Mender and non-Mender builds in this distro, we *do not*
expand the U-Boot environment size and move it around, as
we done in the layout in meta-mender-tegra. Plus we use
config fragments to enable the redundant environment when
using other A/B update systems, like rauc.

So point over to the meta-rauc dynamic layer area to pick
up the config fragments and add a prepend for building
U-Boot with Mender so the config fragments get processed
before Mender's checks take place.

Signed-off-by: Matt Madison <matt@madison.systems>