kernel-module-signing.bbclass: exclude unstripped (debug) modules
Kernel packaging in OE-Core has been updated recently to leverage
more of the common packaging code, so module stripping now happens
the normal way. This means that the debug-enabled kernel modules
are in the tree when we're preparing the signing request, driving
the tarball size up by about 10x.

Fix this by excluding the .debug directories, so we're just signing
the stripped modules.

Signed-off-by: Matt Madison <matt@madison.systems>