swupdate: rework startup and runtime config
Abandon the janky script/command-line configuration
method in favor of creating a configuration file for
swupdate to use.

Config file is a template run through a generator script
at boot time as an ExecStartPre, which installs it under
/run/swupdate.  /etc/swupdate.cfg points there.

Also add a some service dropins to enable this and to
make sure the ustate variable is set back to 0 on
successful boot.

Signed-off-by: Matt Madison <matt@madison.systems>