tegra-sysinstall-tools: hack for tegra-bootinfo
For an unknown (unknowable?) reason, the first time the autoinstaller
tries to run 'tegra-bootinfo --initialize --force-initialize' we get a
Segmentation Fault on jetson-agx-xavier-devkit.

Add a patch which runs with gdb and strace on failure.
Add a patch which tries again and reports success "[OK - 2nd try]"
Add gdb and strace to EXTRA_RDEPENDS.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>