dmc-passphrase.service: use luks-srv-app
Replace keystoretool with luks-srv-app which communicates with the
luks-srv Trusted Application present in tos-prebuilt.

Drop the check for odm_prod_mode as LUKS encryption on its own is not
dependent on any fuses or keys (assume default null KEK2 and null user_key).

Add ExecStartPost to chmod passphrase to be rw for root only, as
cryptsetup will complain if it is world readable.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>