initrd-crypt-setup: use luks-srv-app
Replace keystore-tools with luks-srv-app.

Introduce DISK_ENCRYPTION_CONTEXT variable which can be used to set the
--context-string passed to luks-srv-app. The context is always required,
so provide a weak default 'dummy-context' value.

DISK_ENCRYPTION_CONTEXT can be set in local.conf. As an example, it can
be populated with uuidgen for a random value. It is written in plain
text in /etc/systemd/system/dmc-passphrase.service so the context is
only marginally valuable.