tegra-sysinstall-tools: WIP patches
Add commented out patch to set -x for debugging the tegra-bootinfo
Segmentation Fault.

Add patch to align a bit closer to what NVidia is passing to cryptsetup,
explicitly setting luks1 and key-size to 512.

Add patch to ignore secure boot enabled. It is not strictly required for
LUKS encryption to be functional (assuming null KEK2 and null user_key).

Signed-off-by: Tim Orling <tim.orling@konsulko.com>